First stab at plan(bigint)

+++ test/sql/moretap.sql
@@ -69,7 +69,7 @@ SELECT * FROM test_variadic();
 -- Check no_plan.
 DELETE FROM __tcache__ WHERE label = 'plan';
 SELECT * FROM no_plan();
-SELECT is( value, 0, 'no_plan() should have stored a plan of 0' )
+SELECT is( value, 0::bigint, 'no_plan() should have stored a plan of 0' )
   FROM __tcache__
  WHERE label = 'plan';
